+++ downloadTemplates.js
@@ -5,9 +5,13 @@ import client from 'https';
 
 const releaseBranch = 'release/20250324';
 
+const shouldDownloadImages = process.argv[2] === '--download-images';
+const subdirectory = process.argv[3] === '--subdirectory' ? process.argv[4] : undefined;
 const baseURL = `https://raw.githubusercontent.com/azure/LogicAppsTemplates/${releaseBranch}`;
 const sourceCodeURL = `https://github.com/Azure/LogicAppsTemplates/tree/${releaseBranch}`;
-const templatesFolder = './libs/logic-apps-shared/src/designer-client-services/lib/templates';
+const templatesFolder = subdirectory
+  ? `./libs/logic-apps-shared/src/designer-client-services/lib/templates/${subdirectory}`
+  : './libs/logic-apps-shared/src/designer-client-services/lib/templates';
 
 const downloadFile = async (path) => {
   const artifactUrl = `${baseURL}/${path}`;
@@ -53,6 +57,11 @@ const downloadWorkflowManifest = async (path) => {
   }
 
   if (workflowManifest.images.light && workflowManifest.images.dark) {
+    if (shouldDownloadImages) {
+      await downloadFile(`${path}/${workflowManifest.images.light}.png`);
+      await downloadFile(`${path}/${workflowManifest.images.dark}.png`);
+    }
+
     workflowManifest.images = {
       light: `${baseURL}/${path}/${workflowManifest.images.light}.png`,
       dark: `${baseURL}/${path}/${workflowManifest.images.dark}.png`,
@@ -66,6 +75,14 @@ const downloadWorkflowManifest = async (path) => {
 };
 
 const run = async () => {
+  if (shouldDownloadImages) {
+    console.log('Templates will be downloaded with images');
+  }
+
+  if (subdirectory) {
+    console.log(`Templates will be downloaded to the subdirectory: ${subdirectory}`);
+  }
+
   await removeTemplatesFolderIfPresent();
   await createTemplatesFolder('');
   const registeredManifestNames = await (await downloadFetchFile('manifest.json')).json();