+++ libs/designer-ui/src/lib/editor/initializevariable/index.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useCallback, useMemo, useState } from 'react';
 import type { BaseEditorProps } from '../base';
 import { createEmptyLiteralValueSegment } from '../base/utils/helper';
 import type { ValueSegment } from '../models/parameter';
@@ -38,11 +38,19 @@ export const InitializeVariableEditor = ({
   ...props
 }: InitializeVariableEditorProps) => {
   const intl = useIntl();
-  const [variables, setVariables] = useState<InitializeVariableProps[]>(() =>
+  const [variables, setVariables] = useState<InitializeVariableProps[] | undefined>(() =>
     isAgentParameter ? parseSchemaAsVariableEditorSegments(initialValue) : parseVariableEditorSegments(initialValue)
   );
 
-  const addVariableLabel = intl.formatMessage({ defaultMessage: 'Add a Variable', id: 'HET2nV', description: 'label to add a variable' });
+  const stringResources = useMemo(
+    () => ({
+      ADD_VARIABLE: intl.formatMessage({ defaultMessage: 'Add a Variable', id: 'HET2nV', description: 'label to add a variable' }),
+      ADD_PARAMETER: intl.formatMessage({ defaultMessage: 'Add a Parameter', id: 'ysSmGO', description: 'label to add a parameter' }),
+    }),
+    [intl]
+  );
+
+  const addButtonText = isAgentParameter ? stringResources.ADD_PARAMETER : stringResources.ADD_VARIABLE;
 
   const warningTitleVariable = intl.formatMessage({
     defaultMessage: 'Unable to Parse Variables',
@@ -68,38 +76,48 @@ export const InitializeVariableEditor = ({
     description: 'Warning body for when unable to parse schema',
   });
 
-  const addVariable = () => {
+  const addVariable = useCallback(() => {
     setVariables((prev) => [
-      ...prev,
+      ...(prev ?? []),
       {
         name: [createEmptyLiteralValueSegment()],
         type: [createEmptyLiteralValueSegment()],
         value: [createEmptyLiteralValueSegment()],
       },
     ]);
-  };
+  }, [setVariables]);
 
-  const updateVariables = (updatedVariables: InitializeVariableProps[]) => {
-    const segments = isAgentParameter
-      ? convertVariableEditorSegmentsAsSchema(updatedVariables)
-      : createVariableEditorSegments(updatedVariables);
-    onChange?.({ value: segments, viewModel: { variables: updatedVariables, hideParameterErrors: true } });
-    return updatedVariables;
-  };
+  const updateVariables = useCallback(
+    (updatedVariables: InitializeVariableProps[]) => {
+      const segments = isAgentParameter
+        ? convertVariableEditorSegmentsAsSchema(updatedVariables)
+        : createVariableEditorSegments(updatedVariables);
+      onChange?.({ value: segments, viewModel: { variables: updatedVariables, hideParameterErrors: true } });
+      return updatedVariables;
+    },
+    [isAgentParameter, onChange]
+  );
 
-  const handleDeleteVariable = (index: number) => {
-    setVariables((prev) => {
-      const updatedVariables = prev.filter((_, i) => i !== index);
-      return updateVariables(updatedVariables);
-    });
-  };
+  const handleDeleteVariable = useCallback(
+    (index: number) => {
+      setVariables((prev) => {
+        const p = prev ?? [];
+        const updatedVariables = p.filter((_, i) => i !== index);
+        return updateVariables(updatedVariables);
+      });
+    },
+    [updateVariables]
+  );
 
-  const handleVariableChange = (value: InitializeVariableProps, index: number) => {
-    setVariables((prev) => {
-      const updatedVariables = prev.map((v, i) => (i === index ? value : v));
-      return updateVariables(updatedVariables);
-    });
-  };
+  const handleVariableChange = useCallback(
+    (value: InitializeVariableProps, index: number) => {
+      setVariables((prev) => {
+        const updatedVariables = (prev ?? []).map((v, i) => (i === index ? value : v));
+        return updateVariables(updatedVariables);
+      });
+    },
+    [updateVariables]
+  );
 
   return variables ? (
     <div className="msla-editor-initialize-variables">
@@ -116,12 +134,11 @@ export const InitializeVariableEditor = ({
           errors={validationErrors?.[index]}
         />
       ))}
-
       {props.isMultiVariableEnabled || isAgentParameter ? (
         <div className="msla-initialize-variable-add-variable-button">
           <Button
             appearance="subtle"
-            aria-label={addVariableLabel}
+            aria-label={addButtonText}
             onClick={addVariable}
             disabled={variables.length === constants.PARAMETER.VARIABLE_EDITOR_MAX_VARIABLES || props.readonly}
             icon={<CreateIcon />}
@@ -131,7 +148,7 @@ export const InitializeVariableEditor = ({
                 : { color: 'var(--colorBrandForeground1)', border: '1px solid #9e9e9e' }
             }
           >
-            {addVariableLabel}
+            {addButtonText}
           </Button>
         </div>
       ) : null}
@@ -139,6 +156,7 @@ export const InitializeVariableEditor = ({
   ) : (
     <>
       <StringEditor {...props} initialValue={initialValue} />
+      {/** TODO: Error should only be shown if there's no error */}
       <MessageBar key={'warning'} intent={'warning'} className="msla-initialize-variable-warning">
         <MessageBarBody>
           <MessageBarTitle>{isAgentParameter ? warningTitleAgentParameter : warningTitleVariable}</MessageBarTitle>