+++ libs/designer/src/lib/core/state/workflow/workflowSelectors.ts
@@ -321,16 +321,27 @@ export const useParentRunIndex = (id: string | undefined): number | undefined =>
       if (!id) {
         return undefined;
       }
-      const parents = getAllParentsForNode(id, state.nodesMetadata).filter((x) => {
-        const operationType = getRecordEntry(state.operations, x)?.type?.toLowerCase();
-        return operationType ? operationType === constants.NODE.TYPE.FOREACH || operationType === constants.NODE.TYPE.UNTIL : false;
+      const allParents = getAllParentsForNode(id, state.nodesMetadata);
+      const parents = allParents.filter((x) => {
+        const operationType = getRecordEntry(state.operations, x)?.type?.toLowerCase() ?? '';
+        return [constants.NODE.TYPE.FOREACH, constants.NODE.TYPE.UNTIL, constants.NODE.TYPE.AGENT].includes(operationType);
       });
       return parents.length ? getRecordEntry(state.nodesMetadata, parents[0])?.runIndex : undefined;
     })
   );
 };
+export const useRunIndex = (id: string | undefined): number | undefined => {
+  return useSelector(
+    createSelector(getWorkflowState, (state: WorkflowState) => {
+      if (!id) {
+        return undefined;
+      }
+      return getRecordEntry(state.nodesMetadata, id)?.runIndex;
+    })
+  );
+};
 
-export const useParentRunId = (id: string | undefined): string | undefined => {
+export const useParentNodeId = (id: string | undefined): string | undefined => {
   return useSelector(
     createSelector(getWorkflowState, (state: WorkflowState) => {
       if (!id) {