+++ libs/designer/src/lib/ui/CustomNodes/ScopeCardNode.tsx
@@ -23,28 +23,34 @@ import {
   useRunData,
   useParentRunIndex,
   useRunInstance,
-  useParentRunId,
+  useParentNodeId,
   useNodeDescription,
   useShouldNodeFocus,
+  useRunIndex,
 } from '../../core/state/workflow/workflowSelectors';
-import { setRepetitionRunData, toggleCollapsedGraphId } from '../../core/state/workflow/workflowSlice';
+import {
+  setRepetitionRunData,
+  toggleCollapsedGraphId,
+  updateAgenticGraph,
+  updateAgenticMetadata,
+} from '../../core/state/workflow/workflowSlice';
 import type { AppDispatch } from '../../core/store';
 import { LoopsPager } from '../common/LoopsPager/LoopsPager';
-import { getRepetitionName } from '../common/LoopsPager/helper';
+import { getRepetitionName, getScopeRepetitionName } from '../common/LoopsPager/helper';
 import { DropZone } from '../connections/dropzone';
 import { MessageBarType } from '@fluentui/react';
-import { RunService, equals, isNullOrUndefined, removeIdTag, useNodeIndex } from '@microsoft/logic-apps-shared';
+import { equals, isNullOrUndefined, removeIdTag, useNodeIndex } from '@microsoft/logic-apps-shared';
 import { ScopeCard } from '@microsoft/designer-ui';
 import type { LogicAppsV2 } from '@microsoft/logic-apps-shared';
 import { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useDrag } from 'react-dnd';
 import { useIntl } from 'react-intl';
-import { useQuery } from '@tanstack/react-query';
 import { useDispatch } from 'react-redux';
 import { Handle, Position, type NodeProps } from '@xyflow/react';
 import { copyScopeOperation } from '../../core/actions/bjsworkflow/copypaste';
 import { useHotkeys } from 'react-hotkeys-hook';
 import { CopyTooltip } from '../common/DesignerContextualMenu/CopyTooltip';
+import { useNodeRepetition, useAgentRepetition } from '../../core/queries/runs';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const ScopeCardNode = ({ data, targetPosition = Position.Top, sourcePosition = Position.Bottom, id }: NodeProps) => {
@@ -63,43 +69,52 @@ const ScopeCardNode = ({ data, targetPosition = Position.Top, sourcePosition = P
   const parentRunIndex = useParentRunIndex(scopeId);
   const runInstance = useRunInstance();
   const runData = useRunData(scopeId);
-  const parentRunId = useParentRunId(scopeId);
-  const parentRunData = useRunData(parentRunId ?? '');
+  const parentNodeId = useParentNodeId(scopeId);
+  const parentRunData = useRunData(parentNodeId ?? '');
   const selfRunData = useRunData(scopeId);
   const nodesMetaData = useNodesMetadata();
+  const isPinned = useIsNodePinnedToOperationPanel(scopeId);
+  const selected = useIsNodeSelectedInOperationPanel(scopeId);
+  const brandColor = useBrandColor(scopeId);
+  const iconUri = useIconUri(scopeId);
+  const isLeaf = useIsLeafNode(id);
+  const label = useNodeDisplayName(scopeId);
+  const normalizedType = node?.type.toLowerCase();
+  const isAgent = normalizedType === constants.NODE.TYPE.AGENT;
+  const runIndex = useRunIndex(scopeId);
+  const scopeRepetitionName = useMemo(() => getScopeRepetitionName(runIndex), [runIndex]);
+
   const repetitionName = useMemo(
     () => getRepetitionName(parentRunIndex, scopeId, nodesMetaData, operationsInfo),
     [nodesMetaData, operationsInfo, parentRunIndex, scopeId]
   );
   const rootRef = useRef<HTMLDivElement | null>(null);
-
-  const { isFetching: isRepetitionFetching, data: repetitionRunData } = useQuery<any>(
-    ['runInstance', { nodeId: scopeId, runId: runInstance?.id, repetitionName, parentStatus: parentRunData?.status }],
-    async () => {
-      if (parentRunData?.status === constants.FLOW_STATUS.SKIPPED) {
-        return {
-          properties: {
-            status: constants.FLOW_STATUS.SKIPPED,
-            inputsLink: null,
-            outputsLink: null,
-            startTime: null,
-            endTime: null,
-            trackingId: null,
-            correlation: null,
-          },
-        };
-      }
-
-      return await RunService().getRepetition({ nodeId: scopeId, runId: runInstance?.id }, repetitionName);
-    },
-    {
-      refetchOnWindowFocus: false,
-      refetchOnReconnect: false,
-      refetchOnMount: false,
-      retryOnMount: false,
-      enabled: parentRunIndex !== undefined && !!isMonitoringView,
-    }
+  const { isFetching: isRepetitionFetching, data: repetitionRunData } = useNodeRepetition(
+    !!isMonitoringView,
+    scopeId,
+    runInstance?.id,
+    repetitionName,
+    parentRunData?.status,
+    parentRunIndex
   );
+
+  const { isFetching: isScopeRepetitionFetching, data: scopeRepetitionRunData } = useAgentRepetition(
+    !!isMonitoringView,
+    isAgent,
+    scopeId,
+    runInstance?.id,
+    scopeRepetitionName,
+    parentRunData?.status,
+    runIndex
+  );
+
+  useEffect(() => {
+    if (!isNullOrUndefined(scopeRepetitionRunData)) {
+      const updatePayload = { nodeId: scopeId, scopeRepetitionRunData: scopeRepetitionRunData.properties } as any;
+      dispatch(updateAgenticGraph(updatePayload));
+      dispatch(updateAgenticMetadata(updatePayload));
+    }
+  }, [dispatch, scopeRepetitionRunData, scopeId, selfRunData?.correlation?.actionTrackingId]);
 
   useEffect(() => {
     if (!isNullOrUndefined(repetitionRunData)) {
@@ -107,7 +122,6 @@ const ScopeCardNode = ({ data, targetPosition = Position.Top, sourcePosition = P
         // if the correlation id is the same, we don't need to update the repetition run data
         return;
       }
-
       dispatch(setRepetitionRunData({ nodeId: scopeId, runData: repetitionRunData.properties as LogicAppsV2.WorkflowRunAction }));
     }
   }, [dispatch, repetitionRunData, scopeId, selfRunData?.correlation?.actionTrackingId]);
@@ -146,13 +160,6 @@ const ScopeCardNode = ({ data, targetPosition = Position.Top, sourcePosition = P
     }),
     [readOnly, metadata]
   );
-
-  const isPinned = useIsNodePinnedToOperationPanel(scopeId);
-  const selected = useIsNodeSelectedInOperationPanel(scopeId);
-  const brandColor = useBrandColor(scopeId);
-  const iconUri = useIconUri(scopeId);
-  const isLeaf = useIsLeafNode(id);
-  const label = useNodeDisplayName(scopeId);
 
   const nodeClick = useCallback(() => {
     dispatch(changePanelNode(scopeId));
@@ -204,8 +211,8 @@ const ScopeCardNode = ({ data, targetPosition = Position.Top, sourcePosition = P
   const opQuery = useOperationQuery(scopeId);
 
   const isLoading = useMemo(
-    () => isRepetitionFetching || opQuery.isLoading || (!brandColor && !iconUri),
-    [brandColor, iconUri, opQuery.isLoading, isRepetitionFetching]
+    () => isRepetitionFetching || isScopeRepetitionFetching || opQuery.isLoading || (!brandColor && !iconUri),
+    [brandColor, iconUri, opQuery.isLoading, isRepetitionFetching, isScopeRepetitionFetching]
   );
 
   const comment = useMemo(
@@ -221,35 +228,62 @@ const ScopeCardNode = ({ data, targetPosition = Position.Top, sourcePosition = P
     [brandColor, nodeComment]
   );
 
-  const opManifestErrorText = intl.formatMessage({
-    defaultMessage: 'Error fetching manifest',
-    id: 'HmcHoE',
-    description: 'Error message when manifest fails to load',
-  });
+  const actionCount = metadata?.actionCount ?? 0;
+
+  const intlText = useMemo(
+    () => ({
+      opManifestErrorText: intl.formatMessage({
+        defaultMessage: 'Error fetching manifest',
+        id: 'HmcHoE',
+        description: 'Error message when manifest fails to load',
+      }),
+      settingValidationErrorText: intl.formatMessage({
+        defaultMessage: 'Invalid settings',
+        id: 'Jil/Wa',
+        description: 'Text to explain that there are invalid settings for this node',
+      }),
+      parameterValidationErrorText: intl.formatMessage({
+        defaultMessage: 'Invalid parameters',
+        id: 'Tmr/9e',
+        description: 'Text to explain that there are invalid parameters for this node',
+      }),
+      actionString: intl.formatMessage(
+        {
+          defaultMessage: '{actionCount, plural, one {# Action} =0 {0 Actions} other {# Actions}}',
+          id: 'B/JzwK',
+          description: 'This is the number of actions to be completed in a group',
+        },
+        { actionCount }
+      ),
+      caseString: intl.formatMessage(
+        {
+          defaultMessage: '{actionCount, plural, one {# Case} =0 {0 Cases} other {# Cases}}',
+          id: 'KX1poC',
+          description: 'This is the number of cases or options the program can take',
+        },
+        { actionCount }
+      ),
+      emptyAgent: intl.formatMessage({
+        defaultMessage: 'No tools were executed ',
+        id: 'ZuQSme',
+        description: 'Text to explain that there are no tools in this agent',
+      }),
+    }),
+    [actionCount, intl]
+  );
 
   const settingValidationErrors = useSettingValidationErrors(scopeId);
-  const settingValidationErrorText = intl.formatMessage({
-    defaultMessage: 'Invalid settings',
-    id: 'Jil/Wa',
-    description: 'Text to explain that there are invalid settings for this node',
-  });
-
   const parameterValidationErrors = useParameterValidationErrors(scopeId);
-  const parameterValidationErrorText = intl.formatMessage({
-    defaultMessage: 'Invalid parameters',
-    id: 'Tmr/9e',
-    description: 'Text to explain that there are invalid parameters for this node',
-  });
 
   const { errorMessage, errorLevel } = useMemo(() => {
     if (opQuery?.isError) {
-      return { errorMessage: opManifestErrorText, errorLevel: MessageBarType.error };
+      return { errorMessage: intlText.opManifestErrorText, errorLevel: MessageBarType.error };
     }
     if (settingValidationErrors?.length > 0) {
-      return { errorMessage: settingValidationErrorText, errorLevel: MessageBarType.severeWarning };
+      return { errorMessage: intlText.settingValidationErrorText, errorLevel: MessageBarType.severeWarning };
     }
     if (parameterValidationErrors?.length > 0) {
-      return { errorMessage: parameterValidationErrorText, errorLevel: MessageBarType.severeWarning };
+      return { errorMessage: intlText.parameterValidationErrorText, errorLevel: MessageBarType.severeWarning };
     }
 
     if (isMonitoringView) {
@@ -258,19 +292,10 @@ const ScopeCardNode = ({ data, targetPosition = Position.Top, sourcePosition = P
     }
 
     return { errorMessage: undefined, errorLevel: undefined };
-  }, [
-    opQuery?.isError,
-    opManifestErrorText,
-    settingValidationErrors?.length,
-    settingValidationErrorText,
-    parameterValidationErrors?.length,
-    parameterValidationErrorText,
-    isMonitoringView,
-    runData,
-  ]);
+  }, [opQuery?.isError, intlText, settingValidationErrors?.length, parameterValidationErrors?.length, isMonitoringView, runData]);
 
   const renderLoopsPager = useMemo(() => {
-    if (metadata?.runData?.status && !equals(metadata.runData.status, 'InProgress')) {
+    if (!Array.isArray(metadata?.runData) && metadata?.runData?.status && !equals(metadata.runData.status, 'InProgress')) {
       return <LoopsPager metadata={metadata} scopeId={scopeId} collapsed={graphCollapsed} />;
     }
     return null;
@@ -282,33 +307,16 @@ const ScopeCardNode = ({ data, targetPosition = Position.Top, sourcePosition = P
     return null;
   }
 
-  const normalizedType = node?.type.toLowerCase();
-  const actionCount = metadata?.actionCount ?? 0;
-
-  const actionString = intl.formatMessage(
-    {
-      defaultMessage: '{actionCount, plural, one {# Action} =0 {0 Actions} other {# Actions}}',
-      id: 'B/JzwK',
-      description: 'This is the number of actions to be completed in a group',
-    },
-    { actionCount }
-  );
-
-  const caseString = intl.formatMessage(
-    {
-      defaultMessage: '{actionCount, plural, one {# Case} =0 {0 Cases} other {# Cases}}',
-      id: 'KX1poC',
-      description: 'This is the number of cases or options the program can take',
-    },
-    { actionCount }
-  );
-
   const collapsedText =
-    normalizedType === constants.NODE.TYPE.SWITCH || normalizedType === constants.NODE.TYPE.IF ? caseString : actionString;
+    normalizedType === constants.NODE.TYPE.SWITCH || normalizedType === constants.NODE.TYPE.IF || isAgent
+      ? intlText.caseString
+      : intlText.actionString;
 
   const isFooter = id.endsWith('#footer');
-  const isAgent = normalizedType === constants.NODE.TYPE.AGENT;
   const showEmptyGraphComponents = isLeaf && !graphCollapsed && !isFooter && !isAgent;
+
+  const shouldShowPager = (normalizedType === constants.NODE.TYPE.FOREACH || isAgent) && isMonitoringView;
+  const isCardActive = isMonitoringView ? !isNullOrUndefined(runData?.status) : true;
 
   return (
     <>
@@ -316,7 +324,8 @@ const ScopeCardNode = ({ data, targetPosition = Position.Top, sourcePosition = P
         <div ref={rootRef}>
           <Handle className="node-handle top" type="target" position={targetPosition} isConnectable={false} />
           <ScopeCard
-            active={isMonitoringView ? !isNullOrUndefined(runData?.status) : true}
+            active={isCardActive}
+            showStatusPill={!isAgent && isMonitoringView && isCardActive}
             brandColor={brandColor}
             icon={iconUri}
             isLoading={isLoading}
@@ -329,7 +338,6 @@ const ScopeCardNode = ({ data, targetPosition = Position.Top, sourcePosition = P
             errorMessage={errorMessage}
             isDragging={isDragging}
             id={scopeId}
-            isMonitoringView={isMonitoringView}
             title={label}
             readOnly={readOnly}
             onClick={nodeClick}
@@ -342,13 +350,18 @@ const ScopeCardNode = ({ data, targetPosition = Position.Top, sourcePosition = P
             nodeIndex={nodeIndex}
           />
           {showCopyCallout ? <CopyTooltip id={scopeId} targetRef={rootRef} hideTooltip={clearCopyCallout} /> : null}
-          {normalizedType === constants.NODE.TYPE.FOREACH && isMonitoringView ? renderLoopsPager : null}
+          {shouldShowPager ? renderLoopsPager : null}
           <Handle className="node-handle bottom" type="source" position={sourcePosition} isConnectable={false} />
         </div>
       </div>
       {graphCollapsed && !isFooter ? (
         <p className="no-actions-text" data-automation-id={`scope-${id}-no-actions`}>
           {collapsedText}
+        </p>
+      ) : null}
+      {isAgent && actionCount === 0 && !graphCollapsed ? (
+        <p className="no-actions-text" data-automation-id={`scope-${id}-no-tools`}>
+          {intlText.emptyAgent}
         </p>
       ) : null}
       {showEmptyGraphComponents ? (