+++ libs/designer/src/lib/ui/panel/nodeDetailsPanel/tabs/parametersTab/index.tsx
@@ -454,7 +454,7 @@ const ParameterSection = ({
     if (!conditionParameter?.id) {
       return;
     }
-    const previousConditionValue = parseSchemaAsVariableEditorSegments(conditionParameter.value);
+    const previousConditionValue = parseSchemaAsVariableEditorSegments(conditionParameter.value) ?? [];
     previousConditionValue.push({
       name: [createLiteralValueSegment(name)],
       type: [createLiteralValueSegment(type)],